refactor(schemas): renombrar labelDescription a fieldHelp en definiciones y refs
