+++ lib/schemas/definitions/index.js
@@ -15,7 +15,7 @@ const redirect = require('../common/redirect');
 const canCreate = require('../common/canCreate');
 const actionField = require('../common/actions/field');
 const width = require('../common/width');
-const labelDescription = require('../common/labelDescription');
+const fieldHelp = require('../common/fieldHelp');
 
 module.exports = {
 	$id: 'schemaDefinitions',
@@ -35,6 +35,6 @@ module.exports = {
 		canCreate,
 		actionField,
 		width,
-		labelDescription
+		fieldHelp
 	}
 };
