+++ lib/schemas/edit-new/modules/field.js
@@ -35,7 +35,7 @@ module.exports = {
 		componentAttributes: {
 			type: 'object',
 			properties: {
-				labelDescription: { $ref: 'schemaDefinitions#/definitions/labelDescription' }
+				fieldHelp: { $ref: 'schemaDefinitions#/definitions/fieldHelp' }
 			},
 			default: {}
 		},
