+++ lib/schemas/utils.js
@@ -17,7 +17,7 @@ const makeComponent = ({ name, properties = {}, conditions = {}, requiredPropert
 		...conditions,
 		type: 'object',
 		properties: {
-			labelDescription: { $ref: 'schemaDefinitions#/definitions/labelDescription' },
+			fieldHelp: { $ref: 'schemaDefinitions#/definitions/fieldHelp' },
 			...properties
 		},
 		default: {},
