+++ tests/mocks/schemas/expected/edit.json
@@ -533,7 +533,7 @@
               "name": "name",
               "component": "Select",
               "componentAttributes": {
-                "labelDescription": "Texto de ayuda visible en el tooltip.",
+                "fieldHelp": "Texto de ayuda visible en el tooltip.",
                 "translateLabels": true,
                 "labelFieldName": "motiveName",
                 "labelPrefix": "common.status.",
